requirements.txt
google-api-python-client>=2.100.0
google-auth>=2.25.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
python-docx>=1.1.0
PyPDF2>=3.0.0
rich>=13.0.0
